Move LZMA mode from the opcode to a flag in the header
LZMA mode signifies the compression of the entire packed data. This mode cannot be applied discontinuously or recursively. There is no need to represent this mode as an opcode in the packed data; having it as a flag in the header is sufficient. The length of the compressed data can be calculated, so there's no need to record it in the packed data.