Delete OpCreateProcess
This opcode is used to launch any script during the unpacking of application files. At this point, unpacking is in progress, and the stub has not released working memory. To launch the script after the stub's working memory is released, OpPostCreateProcess was created in commit (4fee09d8).

This opcode is not in use. There doesn't seem to be any reason to launch any script during unpacking.