Refine script execution procedure
Changed the procedure so that the working directory is only modified when launching the application script if the chdir to the application directory is enabled (--chdir_first). This change was made because there is no point in changing the working directory when either the application cannot be launched or there is no application to launch.
Script information is now deallocated even if the application does not start, as there were cases where memory was allocated for script information when the application did not start.