Refactor error handling in GetImagePath function
In the GetImagePath function, the error handling mechanism was improved. The LAST_ERROR macro is now called before the LocalFree(image_path_w); function. This is necessary as the LAST_ERROR macro calls the GetLastError function and gives error output. Therefore, it needs to be called before any memory is freed using LocalFree in order to accurately obtain and display the error message.