Improvement of Test Debug Mode
Modified the test_debug section of the test. The purpose of this modification is to prevent debug messages from being output to standard error and flowing into the test results when testing the debug mode of the stub. To achieve this, the test was modified to read the standard error output when executing the stub executable file. Additionally, the content of the read standard error is the output in the debug mode of the stub, and a test was added to verify if this content is correct. As a result, the test can correctly verify the output in the debug mode of the stub, and this output will not flow into the test results.