Temporarily remove absolute paths from RUBYOPT environment variable
Added a fix to remove the absolute path to bundler/setup from RUBYOPT during the build process.
This is a temporary measure to ensure compatibility with self-extracting executables built in a bundle exec environment, particularly for Ruby 3.2 and later.
In the future, a more appropriate solution will be implemented.

Fixes #14