Removed the sort_uniq method
Removed the sort_uniq method. This method was originally used when a collection of paths contained a mix of strings and Ocran::Pathname objects. Now, the collection of paths consists solely of Ocran::Pathname objects. Additionally, Ocran::Pathname#eql? (an alias for ==) performs comparisons based on the file system's case sensitivity, allowing the use of the uniq method on the collection of paths.