Improved argument path handling for Ocran.files and error checking
Improved the handling of arg by checking if it exists and is not an empty directory before proceeding. Converted the path to an absolute path to account for potential changes in the working directory during script execution. Updated the logic to handle directories and files appropriately by recursively adding all files in the directory or adding the file itself.

Removed encoding conversion for arg as ARGV encoding is UTF-8 in Windows since Ruby 3.0.
See also: https://bugs.ruby-lang.org/issues/13571