Remove special behavior from Ocran::Pathname#relative_path_from
Ocran::Pathname#relative_path_from had a special behavior compared to Pathname#relative_path_from: when the absolute paths of itself and the argument had different drive letters, it returned the argument's path as is.
However, Ocran's self-extracting executables do not perform installations across different drive letters, and there is no implementation to unify different drive letters before building. It is not an issue if the drive letters of the user's application files and the Ruby environment files are different, but there is no way to determine a single installation destination if the drive letters within each set are different.
This is evident from the fact that the Ocran.find_src_root method treats the result of relative_path_from as an error when it is an absolute path.