Avoid side effects on Ocran.files by using a duplicate variable
Previously, the code was adding and replacing elements directly in Ocran.files, which is a user-specified array of script file paths. This caused side effects and made the code harder to read. Changed to duplicate the value of Ocran.files into a variable and perform operations on this variable instead.