Move ocran.rb from bin to exe directory and update file settings
This commit relocates the ocran.rb command script from the bin directory to the exe directory, aligning with the standard structure for Ruby gems. Additionally, the encoding comment has been removed as UTF-8 is the default setting. Also, `# frozen_string_literal: true` was added to adopt a standard practice in Ruby implementation.